+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
--
+-  Fix `tex` get opinions from the orders on causes page #1410
 
 ## Current